+++ packages/cli/src/index.ts
@@ -1,8 +1,6 @@
 import * as fs from "fs";
 import * as path from "path";
 
-import * as yargs from "yargs";
-
 import {
   CONFIG_MODULE_NAME,
   getProjectConfig,
@@ -14,6 +12,7 @@ import {
   serveSite,
   generateCodeForProject,
   findDuplicateSegmentsInProject,
+  findUsageInProject,
   BuildCLIOptions,
   GenerateCodeCLIOptions,
   TestProjectOptions,
@@ -21,7 +20,14 @@ import {
   restoreProject,
   Dependencies,
   Datasource,
+  benchmarkFeature,
+  showProjectConfig,
+  evaluateFeature,
+  assessDistribution,
+  showProjectInfo,
 } from "@featurevisor/core";
+
+const yargs = require("yargs"); // eslint-disable-line @typescript-eslint/no-var-requires
 
 process.on("unhandledRejection", (reason) => {
   console.error(reason);
@@ -44,13 +50,16 @@ function requireAndGetProjectConfig(rootDirectoryPath) {
   return getProjectConfig(rootDirectoryPath);
 }
 
+const rootDirectoryPath = process.cwd();
+
 async function getDependencies(options): Promise<Dependencies> {
-  const rootDirectoryPath = process.cwd();
-  const projectConfig = requireAndGetProjectConfig(rootDirectoryPath);
-  const datasource = new Datasource(projectConfig, rootDirectoryPath);
+  const useRootDirectoryPath = options.rootDirectoryPath || rootDirectoryPath;
+
+  const projectConfig = requireAndGetProjectConfig(useRootDirectoryPath);
+  const datasource = new Datasource(projectConfig, useRootDirectoryPath);
 
   return {
-    rootDirectoryPath,
+    rootDirectoryPath: useRootDirectoryPath,
     projectConfig,
     datasource,
     options,
@@ -67,8 +76,7 @@ async function main() {
     .command({
       command: "init",
       handler: async function (options) {
-        const deps = await getDependencies(options);
-        const hasError = await initProject(deps.rootDirectoryPath, options.example);
+        const hasError = await initProject(rootDirectoryPath, options.example);
 
         if (hasError) {
           process.exit(1);
@@ -148,7 +156,7 @@ async function main() {
           assertionPattern: options.assertionPattern,
           verbose: options.verbose || false,
           showDatafile: options.showDatafile || false,
-          fast: options.fast || false,
+          onlyFailures: options.onlyFailures || false,
         };
 
         const hasError = await testProject(deps, testOptions);
@@ -222,14 +230,207 @@ async function main() {
         const deps = await getDependencies(options);
 
         try {
-          await findDuplicateSegmentsInProject(deps);
+          await findDuplicateSegmentsInProject(deps, {
+            authors: options.authors,
+          });
         } catch (e) {
           console.error(e.message);
           process.exit(1);
         }
       },
     })
-    .example("$0 generate-code", "generate code from YAMLs")
+    .example("$0 find-duplicate-segments", "list segments with same conditions")
+    .example("$0 find-duplicate-segments --authors", "show the duplicates along with author names")
+
+    /**
+     * Find usage
+     */
+    .command({
+      command: "find-usage",
+      handler: async function (options) {
+        const deps = await getDependencies(options);
+
+        try {
+          await findUsageInProject(deps, {
+            segment: options.segment,
+            attribute: options.attribute,
+
+            unusedSegments: options.unusedSegments,
+            unusedAttributes: options.unusedAttributes,
+          });
+        } catch (e) {
+          console.error(e.message);
+          process.exit(1);
+        }
+      },
+    })
+    .example("$0 find-usage", "find usage of segments and attributes")
+    .example("$0 find-usage --segment=my_segment", "find usage of segment")
+    .example("$0 find-usage --attribute=my_attribute", "find usage of attribute")
+    .example("$0 find-usage --unusedSegments", "find unused segments")
+    .example("$0 find-usage --unusedAttributes", "find unused attributes")
+
+    /**
+     * Benchmark features
+     */
+    .command({
+      command: "benchmark",
+      handler: async function (options) {
+        if (!options.environment) {
+          console.error("Please specify an environment with --environment flag.");
+          process.exit(1);
+        }
+
+        if (!options.feature) {
+          console.error("Please specify a feature with --feature flag.");
+          process.exit(1);
+        }
+
+        const deps = await getDependencies(options);
+
+        try {
+          await benchmarkFeature(deps, {
+            environment: options.environment,
+            feature: options.feature,
+            n: parseInt(options.n, 10) || 1,
+            context: options.context ? JSON.parse(options.context) : {},
+            variation: options.variation || undefined,
+            variable: options.variable || undefined,
+          });
+        } catch (e) {
+          console.error(e.message);
+          process.exit(1);
+        }
+      },
+    })
+    .example("$0 benchmark", "benchmark feature evaluations")
+    .example(
+      "$0 benchmark --environment=production --feature=my_feature --context='{}' -n=100",
+      "benchmark feature flag evaluation",
+    )
+    .example(
+      "$0 benchmark --environment=production --feature=my_feature --context='{}' --variation -n=100",
+      "benchmark feature variation evaluation",
+    )
+    .example(
+      "$0 benchmark --environment=production --feature=my_feature --context='{}' --variable=my_variable_key -n=100",
+      "benchmark feature variable evaluation",
+    )
+
+    /**
+     * Show config
+     */
+    .command({
+      command: "config",
+      handler: async function (options) {
+        const projectConfig = requireAndGetProjectConfig(rootDirectoryPath);
+
+        showProjectConfig(projectConfig, {
+          print: options.print,
+          pretty: options.pretty,
+        });
+      },
+    })
+    .example("$0 config", "show project configuration")
+    .example("$0 config --print", "print project configuration as JSON")
+    .example("$0 config --print --pretty", "print project configuration as prettified JSON")
+
+    /**
+     * Evaluate
+     */
+    .command({
+      command: "evaluate",
+      handler: async function (options) {
+        if (!options.environment) {
+          console.error("Please specify an environment with --environment flag.");
+          process.exit(1);
+        }
+
+        if (!options.feature) {
+          console.error("Please specify a feature with --feature flag.");
+          process.exit(1);
+        }
+
+        const deps = await getDependencies(options);
+
+        try {
+          await evaluateFeature(deps, {
+            environment: options.environment,
+            feature: options.feature,
+            context: options.context ? JSON.parse(options.context) : {},
+            print: options.print,
+            pretty: options.pretty,
+            verbose: options.verbose,
+          });
+        } catch (e) {
+          console.error(e.message);
+          process.exit(1);
+        }
+      },
+    })
+    .example("$0 evaluate", "evaluate a feature along with its variation and variables")
+    .example(
+      "$0 evaluate --environment=production --feature=my_feature --context='{}'",
+      "evaluate a feature against provided context",
+    )
+
+    /**
+     * Assess distribution
+     */
+    .command({
+      command: "assess-distribution",
+      handler: async function (options) {
+        if (!options.environment) {
+          console.error("Please specify an environment with --environment flag.");
+          process.exit(1);
+        }
+
+        if (!options.feature) {
+          console.error("Please specify a feature with --feature flag.");
+          process.exit(1);
+        }
+
+        const deps = await getDependencies(options);
+
+        try {
+          await assessDistribution(deps, {
+            environment: options.environment,
+            feature: options.feature,
+            context: options.context ? JSON.parse(options.context) : {},
+            populateUuid: Array.isArray(options.populateUuid)
+              ? options.populateUuid
+              : [options.populateUuid as string].filter(Boolean),
+            n: parseInt(options.n, 10) || 1,
+            verbose: options.verbose,
+          });
+        } catch (e) {
+          console.error(e.message);
+          process.exit(1);
+        }
+      },
+    })
+    .example("$0 assess-distribution", "test traffic distribution of a feature")
+    .example(
+      "$0 assess-distribution --environment=production --feature=my_feature --context='{}' --populateUuid=userId --n=100",
+      "test traffic distribution a feature against provided context",
+    )
+
+    /**
+     * Info
+     */
+    .command({
+      command: "info",
+      handler: async function (options) {
+        const deps = await getDependencies(options);
+
+        try {
+          await showProjectInfo(deps);
+        } catch (e) {
+          console.error(e.message);
+          process.exit(1);
+        }
+      },
+    })
 
     /**
      * Options