+++ packages/core/src/tester/testFeature.ts
@@ -9,38 +9,10 @@ import { createInstance, MAX_BUCKETED_NUMBER } from "@featurevisor/sdk";
 
 import { Datasource } from "../datasource";
 import { ProjectConfig } from "../config";
-import { buildDatafile } from "../builder";
-import { SCHEMA_VERSION } from "../config";
 
 import { checkIfArraysAreEqual } from "./checkIfArraysAreEqual";
 import { checkIfObjectsAreEqual } from "./checkIfObjectsAreEqual";
 import { getFeatureAssertionsFromMatrix } from "./matrix";
-
-export async function getDatafileForFeature(
-  featureKey: string,
-  environment: string,
-  projectConfig: ProjectConfig,
-  datasource: Datasource,
-  revision: string = "testing",
-): Promise<DatafileContent> {
-  const requiredChain = await datasource.getRequiredFeaturesChain(featureKey);
-  const featuresToInclude = Array.from(requiredChain);
-
-  const existingState = await datasource.readState(environment);
-  const datafileContent = await buildDatafile(
-    projectConfig,
-    datasource,
-    {
-      schemaVersion: SCHEMA_VERSION,
-      revision,
-      environment: environment,
-      features: featuresToInclude,
-    },
-    existingState,
-  );
-
-  return datafileContent;
-}
 
 export async function testFeature(
   datasource: Datasource,
@@ -82,15 +54,7 @@ export async function testFeature(
         continue;
       }
 
-      const datafileContent =
-        typeof datafileContentByEnvironment[assertion.environment] !== "undefined"
-          ? datafileContentByEnvironment[assertion.environment]
-          : await getDatafileForFeature(
-              test.feature,
-              assertion.environment,
-              projectConfig,
-              datasource,
-            );
+      const datafileContent = datafileContentByEnvironment[assertion.environment];
 
       if (options.showDatafile) {
         console.log("");