+++ packages/core/src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from "./extractKeys";
 export * from "./git";
+export * from "./pretty";